refactoring services dir
