+++ src/pages/Client/Order/index.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 
-import api from '../../../service/api';
+import api from '../../../services/api';
 import { formatPrice } from '../../../util/format';
 
 import { Container, Total, OrderContainer, BackButton } from './styles';
